refactor(tooling): extract scripts/types.ts, drop the sibling import
check-catalog.core.ts owned two things: the catalog rules, and the
package.json vocabulary every check needs. That made check-pack.core.ts --
a module whose header promises no IO and no pnpm -- import the catalog
checker at runtime just to iterate DEP_FIELDS, and made workspace.ts name
its own return type by reaching into a leaf checker.

types.ts now holds DEP_FIELDS, DepField, DependencyMap, Manifest, Member and
Report, and depends on nothing. Verified with an ESM resolve hook: importing
check-pack.core.ts loads only itself and types.ts, and workspace.ts still
loads only itself, since Member erases.

DEP_FIELDS is the one value in a types module because DepField derives from
it; a hand-written union would drift from the array both checks iterate.
That shared array had no test -- dropping optionalDependencies from it left
test:root green while both checks silently stopped inspecting the field.
check-pack.test.ts's "flags catalog: refs in any dependency field" now
covers all four, so dropping any one fails.
